rsi, hurst, connors rsi

+++ src/a-d/ConnorsRsi/ConnorsRsi.Models.cs
@@ -5,10 +5,10 @@ namespace Skender.Stock.Indicators
     [Serializable]
     public class ConnorsRsiResult : ResultBase
     {
-        public decimal? RsiClose { get; set; }
-        public decimal? RsiStreak { get; set; }
-        public decimal? PercentRank { get; set; }
-        public decimal? ConnorsRsi { get; set; }
+        public double? RsiClose { get; set; }
+        public double? RsiStreak { get; set; }
+        public double? PercentRank { get; set; }
+        public double? ConnorsRsi { get; set; }
 
         // internal use only
         internal int? Streak { get; set; }
